Removing Feature which have 70% garbage values
This code can be used for feature selection process. It will remove those columns which have more than 70% of values Null or some garbage values. As these columns will not contribute much in the model creation. To use this, first we need to understand the syntax of the null values present in the data e.g. 9999 in this data. Then we compare this value in the loop.