Ranking Features based on the importance using Boruta
Boruta gives the ranks to the variables based on the importance of these variables. To check the importance of a variable, It iterates the values with in that column and monitors the effect of this variable on the target column. If changing values of this column is affecting the target variable too much then definitely this column is important in target prediction. If changing values of the column is not affecting the target variable then target variable is independent of this variable. Boruta divides all the columns into 3 groups: Very important variables, Not important variables , rest all the variables are listed in the group as 'Can't be decided'.