Handling class imbalance using Smote function
While dealing with real life data, we mostly encounter class imbalance. Class imbalance occurs when each class does not make up an equal portion of your data-set. Suppose, 90% of the classes are Yes and 10% of the classes are No. In this case, if we create a model which always predicts 'Yes', our model will have 90% accuracy which is logically wrong. Then we use smote function to handle class imbalance. It generates new instances in the data so that we have equal proportions of all the classes in the data.