Regularization using Liblinear
Regularization helps us to penalize the coefficients in model creation. It uses a tuning parameter and based on that it decreases the coefficients of the variables to decrease the variance. Ridge regression keeps all the variables in the model while Lasso makes the coefficients of some variables as zero. As we increase the values of tuning parameter, the coefficients tend to decrease.